Rename encoder, write docstrings
